refactor tmx-generator
